feat(inactive-workplace-db): Move findInactiveWorkplaces to models folder
